# docs/swagger.yaml
basePath: /
definitions:
  github_com_desen_track-me_internal_dto.APIResponse:
    properties:
      data: {}
      errors:
        items:
          type: string
        type: array
      message:
        type: string
      success:
        type: boolean
    type: object
...
  github_com_desen_track-me_internal_dto.LocationResponse:
    properties:
      accuracy:
        type: number
      activity:
        type: string
      altitude:
        type: number
      battery:
        type: integer
      bearing:
        type: number
      charging:
        type: boolean
      gps_provider:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      mock_location:
        type: boolean
      speed:
        type: number
      timestamp:
        type: string
      user_id:
        type: string
    type: object
  github_com_desen_track-me_internal_dto.LocationUpdateRequest:
    properties:
      accuracy:
        type: number
      activity:
        type: string
      altitude:
        type: number
      battery:
        type: integer
      bearing:
        type: number
      charging:
        type: boolean
      gps_provider:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      mock_location:
        type: boolean
      speed:
        type: number
    required:
    - latitude
    - longitude
    type: object
...
  github_com_desen_track-me_internal_dto.SharingApproval:
    properties:
      approve:
        type: boolean
      notification_id:
        type: string
    required:
    - notification_id
    type: object
  github_com_desen_track-me_internal_dto.SharingRequest:
    properties:
      friend_id:
        type: string
    required:
    - friend_id
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Backend REST & WebSocket API for lightweight, privacy-focused mutual
    location sharing.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Mutual Location Sharing API
  version: "1.0"
paths:
...
  /api/v1/location/current/{userId}:
    get:
      consumes:
      - application/json
      description: Retrieve the latest GPS location of a friend if mutual sharing
        is active
      parameters:
      - description: Friend User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_desen_track-me_internal_dto.LocationResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
        "401":
          description: Unauthorized or sharing not enabled
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Get friend's current location
      tags:
      - Location
  /api/v1/location/history/{userId}:
    get:
      consumes:
      - application/json
      description: Retrieve location history of a friend if mutual sharing is active
      parameters:
      - description: Friend User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Start timestamp in ISO 8601 / RFC 3339 format
        in: query
        name: start
        required: true
        type: string
      - description: End timestamp in ISO 8601 / RFC 3339 format
        in: query
        name: end
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_desen_track-me_internal_dto.LocationResponse'
                  type: array
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
        "401":
          description: Unauthorized or sharing not enabled
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Get friend's location history
      tags:
      - Location
  /api/v1/location/update:
    post:
      consumes:
      - application/json
      description: Update user's current GPS location and broadcast it to active sharing
        friends
      parameters:
      - description: GPS data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_desen_track-me_internal_dto.LocationUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Location updated
          schema:
            allOf:
            - $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_desen_track-me_internal_dto.LocationResponse'
              type: object
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Post location update
      tags:
      - Location
...
  /api/v1/sharing/approve:
    post:
      consumes:
      - application/json
      description: Approve or deny a pending location sharing request
      parameters:
      - description: Sharing approval details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_desen_track-me_internal_dto.SharingApproval'
      produces:
      - application/json
      responses:
        "200":
          description: Approved or Denied successfully
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Approve location sharing request
      tags:
      - Location
  /api/v1/sharing/request:
    post:
      consumes:
      - application/json
      description: Send a location sharing request to a friend
      parameters:
      - description: Sharing request details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_desen_track-me_internal_dto.SharingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Request sent successfully
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Request location sharing
      tags:
      - Location
  /api/v1/sharing/revoke/{id}:
    delete:
      consumes:
      - application/json
      description: Stop sharing location with a friend
      parameters:
      - description: Friend User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sharing revoked successfully
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_desen_track-me_internal_dto.APIResponse'
      security:
      - BearerAuth: []
      summary: Revoke location sharing
      tags:
      - Location
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token (e.g. Bearer <token>)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"